feat/add items list in discord message
